refactor: split gen-fixtures specs into separate file and add request-scoped clock
